requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
httpx>=0.25.0
google-cloud-aiplatform>=1.36.0
vertexai>=0.0.1
beautifulsoup4>=4.12.0
requests>=2.30.0